Refactor: PostDetailPage 마크업 분리
