+++ src/views/Post/PostDetailPage/Content/index.js
@@ -1,0 +1,2 @@
+export { default as Post } from './Post/Post';
+export { default as Comment } from './Comment/Comment';
